project/requirements.txt
Django==4.2.7
psycopg2==2.9.9
Pillow==10.1.0
pillow-avif-plugin==1.4.6
django-tinymce==4.1.0
django-markdownx
gunicorn==22.0.0
requests==2.32.4
